requirements.txt
asgiref==3.2.5
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.9
django-auditlog==0.4.7
django-bootstrap-form==3.4
django-crispy-forms==1.9.0
django-import-export==2.0.2
django-jsonfield==1.4.0
django-model-utils==4.0.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.2.0
python-dateutil==2.6.0
python-decouple==3.3
pytz==2019.3
PyYAML==5.3
six==1.12.0
sqlparse==0.3.1
static3==0.7.0
tablib==1.1.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
